+++ CHANGES.md
@@ -7,6 +7,7 @@ Changes in 11.7.2
 - Fix Security issue #346 SQL injection in AttendanceSummary.php & Configuration.php, thanks to @fishinspace
 - Fix Security issue #347 SQL injection in index.php & Configuration.php, thanks to @fishinspace
 - Fix potential authenticated SQL injection (administrator), cast to float in Reminders.php
+- Fix authenticated SQL injection (administrator) via Configuration, escape string in RegistrationSave.fnc.php
 
 Changes in 11.7.1
 -----------------